backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hybrid-nl2sql-backend"
version = "0.1.0"
description = "Starter FastAPI backend for OCI GenAI + Spark NL2SQL."
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic-settings>=2.4,<3.0",
  "langchain>=0.3,<0.4",
  "langchain-oci>=0.1.0",
  "pyspark>=3.5,<4.0",
  "python-dotenv>=1.0,<2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
